overlay: capture initial location and size
Do not wait for the next location/size change. Relocate and resize the
overlay immediately. This works on the possibly faulty assumption that
when a process has multiple windows, the largest window should be used.

Also be sure to close the handle to the underlay by expanding Target
to express cleanup operations.